[game] board refactor - moved to /match
