+++ client/src/App.jsx
@@ -5,11 +5,10 @@ import LandingPage from "./components/LandingPage/index.jsx";
 import SocketTest from "./components/LandingPage/socketTest.jsx";
 import Signup from "./components/Auth/Signup.jsx";
 import Login from "./components/Auth/Login.jsx";
-import ShogiBoard from "./components/GameBoard/ShogiBoard.jsx";
 import Home from "./components/Home/index.jsx";
 import Protected from "./components/Global/Protected.jsx";
 import Account from "./components/Account/index.jsx";
-import BoardIndex from "./components/GameBoard/index.jsx";
+import BoardIndex from "./components/Match/index.jsx";
 import Friends from "./components/Friends/index.jsx";
 import EditProfile from "./components/Account/EditProfile.jsx";
 import FrequentlyAsked from "./components/Account/FrequentlyAsked.jsx";
