+++ client/src/components/Match/Match.css
@@ -1,0 +1,200 @@
+/*
+pale red: f15a4a
+light tan: d0c7af
+mid tan: b0a091
+dark tan: 4d443b
+light blue: 485562
+mid blue: 2c343f
+dark blue: 14151b
+*/
+
+@import url('https://fonts.googleapis.com/css?family=Open+Sans:400,700|Oswald');
+
+body {
+  font-size: 14px;
+  font-family: 'Open Sans';
+  font-weight: 400;
+  margin: 0px;
+}
+
+h1, h2, h3, h4, h5 { font-family: 'Oswald' }
+strong { font-weight: 700; }
+
+a {
+  transition: all 0.4s ease;
+}
+
+button {
+  width: 120px;
+  height: 36px;
+  margin: 2px;
+  font-size: 14px;
+  font-weight: 700;
+  text-transform: uppercase;
+}
+
+.action {
+  background-color: #333;
+  text-shadow: 1px 1px 0px #14151b;
+  transition: background-color 0.4s ease;
+}
+
+.action:hover {
+  cursor: pointer;
+  background-color: #f15a4a;
+}
+
+.match {
+  color: #d0c7af;
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+  display: grid;
+  grid-template-columns: 1fr 1fr 6fr 2fr;
+  margin: 0 auto;
+  background: radial-gradient(#485562, #14151b);
+}
+
+.match__status {
+  width: 200px;
+}
+
+.match__turn {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-around;
+}
+
+.match__play {
+  display: grid;
+  grid-template-rows: 1fr 9fr 1fr;
+  min-width: 540px;
+  max-width: 630px;
+  margin: 0 auto;
+}
+
+.match__chat {
+  display: grid;
+  grid-template-rows: 1fr 8fr 1fr;
+}
+
+.match__hand {
+  display: grid;
+  grid-auto-rows: minmax(60px, auto);
+}
+
+.match__board {
+  display: grid;
+  grid-template-columns: repeat(9, 1fr);
+  grid-auto-columns: minmax(60px, 70px);
+  grid-auto-rows: minmax(60px, 70px);
+  border-collapse: collapse;
+  border: solid 1px rgba(255,255,255,0.6);
+  box-shadow: 0px 0px 6px 2px rgba(20, 21, 27, 0.8);
+  margin: 4px;
+}
+
+.match__actions {
+  display: flex;
+  flex-flow: row nowrap;
+  align-items: center;
+  justify-content: space-around;
+}
+
+.shogi__token {
+  text-indent: -999px;
+  overflow: hidden;
+  width: 80%;
+  height: 90%;
+  margin: 0 auto;
+  background-repeat: no-repeat;
+  background-size: contain;
+}
+
+.player__hand-tile > .shogi__token {
+  width: 55%;
+  height: 60%;
+}
+
+.space {
+  border: solid 1px rgba(255,255,255,0.2);
+  background: #d0c7af;
+  margin: 0px;
+  padding: 8px 8px;
+}
+
+.active {
+  cursor: pointer;
+  transition: box-shadow 0.4s ease;
+  box-shadow: inset 0px 0px 0px 0px rgba(241, 90, 74, 0.0);
+}
+
+.active:hover {
+  box-shadow: inset 0px 0px 0px 4px rgba(241, 90, 74, 0.8);
+}
+
+.selected {
+  box-shadow: inset 0px 0px 0px 4px rgba(0,0,0,0.1);
+  background-image: radial-gradient(#ec796c, #f15a4a);
+}
+
+.hinted {
+  box-shadow: inset 0px 0px 0px 4px rgba(241, 90, 74, 0.8);
+  background-image: radial-gradient(#fff, #d0c7af);
+}
+.hinted:hover {
+  cursor: pointer;
+  box-shadow: inset 0px 0px 0px 4px rgba(241, 90, 74, 0.8);
+  background-image: radial-gradient(#fff, #fff);
+}
+
+.player__hand {
+  display: grid;
+  grid-template-columns: repeat(9, 1fr);
+  grid-auto-rows: minmax(60px, auto);
+}
+.north{
+  flex-direction: column-reverse;
+  align-content: flex-end;
+  align-items: flex-end;
+}
+.south{
+  align-content: flex-start;
+  align-items: flex-start;
+}
+
+.player__hand-tile {
+  cursor: pointer;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  justify-content: space-evenly;
+}
+.player__hand-tile.north {
+  margin-top: 4px;
+  flex-direction: column;
+}
+.player__hand-tile.south {
+  margin-bottom: 4px;
+  flex-direction: column-reverse;
+}
+
+.player__hand-count {
+  text-align: center;
+}
+
+.lower-border {
+  border-bottom-color: rgba(0,0,0,0.6);
+}
+.upper-border {
+  border-top-color: rgba(0,0,0,0.6);
+}
+.right-border {
+  border-right-color: rgba(0,0,0,0.6);
+}
+.left-border {
+  border-left-color: rgba(0,0,0,0.6);
+}
