+++ client/src/components/Match/Match.jsx
@@ -6,12 +6,12 @@ import {
   validDropLocations,
   copyMatrix,
   reverseBoard,
-  playerColorFromId,
-  gameTileAtCoords,
-  findKings
+  findKings,
+  gameTileAtCoords
 } from '../../../lib/boardHelpers';
 import GameTile from '../../../lib/GameTile';
 
+import ShogiBoard from '../GameBoard/ShogiBoard.jsx';
 import PlayerPanel from './PlayerPanel.jsx';
 import PlayerHand from './PlayerHand.jsx';
 import MatchLog from './MatchLog.jsx';
@@ -19,7 +19,7 @@ import TurnIndicator from './TurnIndicator.jsx';
 import GameChat from './Chat/index.jsx';
 import ModalPrompt from '../Global/ModalPrompt.jsx';
 
-import './ShogiBoard.css';
+import './Match.css';
 
 class Match extends Component {
   constructor(props) {
@@ -90,6 +90,34 @@ class Match extends Component {
     let localUser = localStorage.getItem('username');
     let updatePlayer = { ...this.state.player };
     let updateOpponent = { ...this.state.opponent };
+
+    if (localUser === this.props.match.black.username) {
+      updatePlayer = {
+        user: this.props.match.black,
+        color: 'black',
+        facing: 'north',
+        hand: this.props.match.hand_black || [],
+      };
+      updateOpponent = {
+        user: this.props.match.white,
+        color: 'white',
+        facing: 'south',
+        hand: this.props.match.hand_white || [],
+      }
+    } else {
+      updatePlayer = {
+        user: this.props.match.white,
+        color: 'white',
+        facing: 'north',
+        hand: this.props.match.hand_white || [],
+      };
+      updateOpponent = {
+        user: this.props.match.black,
+        color: 'black',
+        facing: 'south',
+        hand: this.props.match.hand_black || [],
+      }
+    }
     let updateBoard = updatePlayer.color === 'black' ? reverseBoard(this.state.board) : copyMatrix(this.state.board);
     let kingPositions = findKings(this.state.board, updatePlayer.color);
     let isTurn = this.props.match.turn ? updatePlayer.color === 'black' : updatePlayer.color === 'white';
@@ -397,10 +425,40 @@ class Match extends Component {
           <TurnIndicator isTurn={this.state.isTurn} />
           <PlayerPanel player={this.state.player} />
         </div>
-        <div className="match__play">
-          <PlayerHand location="north" />
-          <ShogiBoard />
-          <PlayerHand location="south" />
+        <div>
+          <div className="match__play">
+            <div className="match__hand">
+              <PlayerHand
+                id={'opponent'}
+                local={false}
+                selected={this.state.selected}
+                player={this.state.opponent}
+                turn={!this.state.isTurn}
+                activate={this.togglePiece}
+              />
+            </div>
+            <div className="match__board">
+              <ShogiBoard
+                board={this.state.board}
+                selected={this.state.selected}
+                hints={this.state.hints}
+                player={this.state.player}
+                isTurn={this.state.isTurn}
+                togglePiece={this.togglePiece}
+                movePiece={this.movePiece}
+              />
+            </div>
+            <div className="match__hand">
+              <PlayerHand
+                id={'player'}
+                local={true}
+                selected={this.state.selected}
+                player={this.state.player}
+                turn={this.state.isTurn}
+                activate={this.togglePiece}
+              />
+            </div>
+          </div>
         </div>
         <div className="match__chat">
           <div></div>
